Refactor desktop UI ownership
