+++ erun-ui/app.go
@@ -65,184 +65,6 @@ type App struct {
 	busyEnvs   map[string]int
 }
 
-type uiState struct {
-	Tenants            []uiTenant     `json:"tenants"`
-	Selected           *uiSelection   `json:"selected,omitempty"`
-	Message            string         `json:"message,omitempty"`
-	Build              uiBuildDetails `json:"build"`
-	VersionSuggestions []uiVersion    `json:"versionSuggestions,omitempty"`
-}
-
-type uiTenant struct {
-	Name         string          `json:"name"`
-	Environments []uiEnvironment `json:"environments"`
-}
-
-type uiEnvironment struct {
-	Name           string `json:"name"`
-	MCPURL         string `json:"mcpUrl,omitempty"`
-	RuntimeVersion string `json:"runtimeVersion,omitempty"`
-	IsActive       bool   `json:"isActive,omitempty"`
-	SSHDEnabled    bool   `json:"sshdEnabled,omitempty"`
-}
-
-type uiSelection struct {
-	Tenant            string `json:"tenant"`
-	Environment       string `json:"environment"`
-	Version           string `json:"version,omitempty"`
-	RuntimeImage      string `json:"runtimeImage,omitempty"`
-	KubernetesContext string `json:"kubernetesContext,omitempty"`
-	ContainerRegistry string `json:"containerRegistry,omitempty"`
-	NoGit             bool   `json:"noGit,omitempty"`
-	Bootstrap         bool   `json:"bootstrap,omitempty"`
-	SetDefaultTenant  bool   `json:"setDefaultTenant,omitempty"`
-	Action            string `json:"action,omitempty"`
-	Debug             bool   `json:"debug,omitempty"`
-}
-
-type uiBuildDetails struct {
-	Version string `json:"version"`
-	Commit  string `json:"commit,omitempty"`
-	Date    string `json:"date,omitempty"`
-}
-
-type uiVersion = eruncommon.RuntimeVersionSuggestion
-
-type uiERunConfig struct {
-	DefaultTenant  string                  `json:"defaultTenant"`
-	CloudProviders []uiCloudProviderStatus `json:"cloudProviders,omitempty"`
-	CloudContexts  []uiCloudContextStatus  `json:"cloudContexts,omitempty"`
-}
-
-type uiTenantConfig struct {
-	Name               string `json:"name"`
-	DefaultEnvironment string `json:"defaultEnvironment"`
-}
-
-type uiSSHDConfig struct {
-	Enabled       bool   `json:"enabled"`
-	LocalPort     int    `json:"localPort"`
-	PublicKeyPath string `json:"publicKeyPath"`
-}
-
-type uiEnvironmentLocalPorts struct {
-	RangeStart int          `json:"rangeStart"`
-	RangeEnd   int          `json:"rangeEnd"`
-	MCP        int          `json:"mcp"`
-	SSH        int          `json:"ssh"`
-	MCPStatus  uiPortStatus `json:"mcpStatus"`
-	SSHStatus  uiPortStatus `json:"sshStatus"`
-}
-
-type uiPortStatus struct {
-	Available bool   `json:"available"`
-	Status    string `json:"status"`
-}
-
-type uiEnvironmentConfig struct {
-	Name                 string                  `json:"name"`
-	RepoPath             string                  `json:"repoPath"`
-	KubernetesContext    string                  `json:"kubernetesContext"`
-	ContainerRegistry    string                  `json:"containerRegistry"`
-	CloudProviderAlias   string                  `json:"cloudProviderAlias"`
-	CloudProviderAliases []string                `json:"cloudProviderAliases,omitempty"`
-	CloudContext         *uiCloudContextStatus   `json:"cloudContext,omitempty"`
-	RuntimeVersion       string                  `json:"runtimeVersion"`
-	SSHD                 uiSSHDConfig            `json:"sshd"`
-	Idle                 uiIdleConfig            `json:"idle"`
-	LocalPorts           uiEnvironmentLocalPorts `json:"localPorts"`
-	Remote               bool                    `json:"remote"`
-	Snapshot             bool                    `json:"snapshot"`
-}
-
-type uiIdleConfig struct {
-	Timeout          string `json:"timeout"`
-	WorkingHours     string `json:"workingHours"`
-	IdleTrafficBytes int64  `json:"idleTrafficBytes"`
-}
-
-type uiCloudProviderStatus struct {
-	Alias     string `json:"alias"`
-	Provider  string `json:"provider"`
-	Username  string `json:"username,omitempty"`
-	AccountID string `json:"accountId,omitempty"`
-	Profile   string `json:"profile,omitempty"`
-	Status    string `json:"status"`
-	Message   string `json:"message,omitempty"`
-}
-
-type uiCloudContextStatus struct {
-	Name               string `json:"name"`
-	Provider           string `json:"provider"`
-	CloudProviderAlias string `json:"cloudProviderAlias"`
-	Region             string `json:"region"`
-	InstanceID         string `json:"instanceId,omitempty"`
-	PublicIP           string `json:"publicIp,omitempty"`
-	InstanceType       string `json:"instanceType"`
-	DiskType           string `json:"diskType"`
-	DiskSizeGB         int    `json:"diskSizeGb"`
-	KubernetesContext  string `json:"kubernetesContext"`
-	Status             string `json:"status"`
-	Message            string `json:"message,omitempty"`
-}
-
-type uiAWSCloudAliasInput struct {
-	Alias       string `json:"alias,omitempty"`
-	Username    string `json:"username,omitempty"`
-	AccountID   string `json:"accountId,omitempty"`
-	Profile     string `json:"profile,omitempty"`
-	SSORegion   string `json:"ssoRegion,omitempty"`
-	SSOStartURL string `json:"ssoStartUrl,omitempty"`
-}
-
-type uiCloudContextInitInput struct {
-	Name               string `json:"name,omitempty"`
-	CloudProviderAlias string `json:"cloudProviderAlias"`
-	Region             string `json:"region"`
-	InstanceType       string `json:"instanceType"`
-	DiskType           string `json:"diskType"`
-	DiskSizeGB         int    `json:"diskSizeGb"`
-}
-
-type startSessionResult struct {
-	SessionID int         `json:"sessionId"`
-	Selection uiSelection `json:"selection"`
-}
-
-type deleteEnvironmentResult struct {
-	Tenant                string `json:"tenant"`
-	Environment           string `json:"environment"`
-	Namespace             string `json:"namespace,omitempty"`
-	KubernetesContext     string `json:"kubernetesContext,omitempty"`
-	NamespaceDeleteError  string `json:"namespaceDeleteError,omitempty"`
-	CloudContextStopError string `json:"cloudContextStopError,omitempty"`
-}
-
-type terminalOutputPayload struct {
-	SessionID int    `json:"sessionId"`
-	Data      string `json:"data"`
-}
-
-type terminalExitPayload struct {
-	SessionID int    `json:"sessionId"`
-	Reason    string `json:"reason,omitempty"`
-}
-
-type appStatusPayload struct {
-	Message string `json:"message"`
-	Busy    bool   `json:"busy"`
-}
-
-type pastedImagePayload struct {
-	Data     string `json:"data"`
-	MIMEType string `json:"mimeType,omitempty"`
-	Name     string `json:"name,omitempty"`
-}
-
-type pastedImageResult struct {
-	Path string `json:"path"`
-}
-
 func NewApp(deps erunUIDeps) *App {
 	deps = withDefaultCoreDeps(deps)
 	deps = withDefaultRuntimeDeps(deps)
@@ -1346,242 +1168,6 @@ func (a *App) LoadDiff(selection uiSelection) (eruncommon.DiffResult, error) {
 		return eruncommon.DiffResult{}, err
 	}
 	return a.deps.loadDiff(ctx, endpoint)
-}
-
-type uiIdleStatus struct {
-	TimeoutSeconds      int64          `json:"timeoutSeconds"`
-	SecondsUntilStop    int64          `json:"secondsUntilStop"`
-	StopEligible        bool           `json:"stopEligible"`
-	OutsideWorkingHours bool           `json:"outsideWorkingHours"`
-	ManagedCloud        bool           `json:"managedCloud"`
-	StopBlockedReason   string         `json:"stopBlockedReason,omitempty"`
-	StopError           string         `json:"stopError,omitempty"`
-	CloudContextName    string         `json:"cloudContextName,omitempty"`
-	CloudContextStatus  string         `json:"cloudContextStatus,omitempty"`
-	CloudContextLabel   string         `json:"cloudContextLabel,omitempty"`
-	Markers             []uiIdleMarker `json:"markers,omitempty"`
-}
-
-type uiIdleMarker struct {
-	Name             string `json:"name"`
-	Idle             bool   `json:"idle"`
-	Reason           string `json:"reason,omitempty"`
-	SecondsRemaining int64  `json:"secondsRemaining,omitempty"`
-}
-
-func (a *App) LoadIdleStatus(selection uiSelection) (uiIdleStatus, error) {
-	selection = normalizeSelection(selection)
-	if selection.Tenant == "" || selection.Environment == "" {
-		return uiIdleStatus{}, fmt.Errorf("tenant and environment are required")
-	}
-	result, err := eruncommon.ResolveOpen(a.deps.store, eruncommon.OpenParams{
-		Tenant:      selection.Tenant,
-		Environment: selection.Environment,
-	})
-	if err != nil {
-		return uiIdleStatus{}, err
-	}
-	mcpPort := eruncommon.MCPPortForResult(result)
-	if !a.deps.canConnectLocalPort(mcpPort) {
-		status, err := a.loadLocalIdleStatus(result)
-		if err == nil {
-			a.maybeStopIdleCloudEnvironment(result, status.status)
-		}
-		return status.ui, err
-	}
-	ctx := a.ctx
-	if ctx == nil {
-		ctx = context.Background()
-	}
-	status, err := a.deps.loadIdleStatus(ctx, mcpEndpointForOpenResult(result))
-	if err != nil {
-		status, err := a.loadLocalIdleStatus(result)
-		if err == nil {
-			a.maybeStopIdleCloudEnvironment(result, status.status)
-		}
-		return status.ui, err
-	}
-	merged := a.mergeLocalIdleActivity(result, status)
-	a.maybeStopIdleCloudEnvironment(result, merged)
-	return a.idleStatusToUI(result, merged), nil
-}
-
-type resolvedUIIdleStatus struct {
-	ui     uiIdleStatus
-	status eruncommon.EnvironmentIdleStatus
-}
-
-func (a *App) loadLocalIdleStatus(result eruncommon.OpenResult) (resolvedUIIdleStatus, error) {
-	status, err := eruncommon.ResolveStoredEnvironmentIdleStatus(a.deps.store, result.Tenant, result.Environment, time.Now())
-	if err != nil {
-		return resolvedUIIdleStatus{}, err
-	}
-	return resolvedUIIdleStatus{ui: a.idleStatusToUI(result, status), status: status}, nil
-}
-
-func (a *App) idleStatusToUI(result eruncommon.OpenResult, status eruncommon.EnvironmentIdleStatus) uiIdleStatus {
-	ui := idleStatusToUI(status)
-	cloudContext, ok, err := a.linkedCloudContext(result.EnvConfig)
-	if err != nil || !ok {
-		return ui
-	}
-	ui.CloudContextName = strings.TrimSpace(cloudContext.Name)
-	ui.CloudContextStatus = strings.TrimSpace(cloudContext.Status)
-	ui.CloudContextLabel = cloudContextDisplayName(cloudContext)
-	return ui
-}
-
-func idleStatusToUI(status eruncommon.EnvironmentIdleStatus) uiIdleStatus {
-	markers := make([]uiIdleMarker, 0, len(status.Markers))
-	for _, marker := range status.Markers {
-		markers = append(markers, uiIdleMarker{
-			Name:             strings.TrimSpace(marker.Name),
-			Idle:             marker.Idle,
-			Reason:           strings.TrimSpace(marker.Reason),
-			SecondsRemaining: marker.SecondsRemaining,
-		})
-	}
-	return uiIdleStatus{
-		TimeoutSeconds:      int64(status.Policy.Timeout / time.Second),
-		SecondsUntilStop:    activitySecondsUntilIdle(status),
-		StopEligible:        status.StopEligible,
-		OutsideWorkingHours: status.OutsideWorkingHours,
-		ManagedCloud:        status.ManagedCloud,
-		StopBlockedReason:   strings.TrimSpace(status.StopBlockedReason),
-		StopError:           strings.TrimSpace(status.StopError),
-		Markers:             markers,
-	}
-}
-
-func (a *App) mergeLocalIdleActivity(result eruncommon.OpenResult, status eruncommon.EnvironmentIdleStatus) eruncommon.EnvironmentIdleStatus {
-	local, err := eruncommon.ResolveStoredEnvironmentIdleStatus(a.deps.store, result.Tenant, result.Environment, time.Now())
-	if err != nil {
-		return status
-	}
-	if len(status.Activity) > 0 {
-		remoteWithLocalPolicy, err := eruncommon.ResolveEnvironmentIdleStatus(result.EnvConfig.Idle, status.Activity, time.Now())
-		if err == nil {
-			remoteWithLocalPolicy.ManagedCloud = status.ManagedCloud
-			remoteWithLocalPolicy.StopBlockedReason = status.StopBlockedReason
-			remoteWithLocalPolicy.StopError = status.StopError
-			status = remoteWithLocalPolicy
-		}
-	}
-	return mergeNewerActivityMarkers(status, local)
-}
-
-func mergeNewerActivityMarkers(status, local eruncommon.EnvironmentIdleStatus) eruncommon.EnvironmentIdleStatus {
-	status.ManagedCloud = local.ManagedCloud
-	status.StopBlockedReason = local.StopBlockedReason
-	for _, localMarker := range local.Markers {
-		if localMarker.Name == "working-hours" || localMarker.LastActivity.IsZero() {
-			continue
-		}
-		for index, marker := range status.Markers {
-			if marker.Name == localMarker.Name && localMarker.LastActivity.After(marker.LastActivity) {
-				status.Markers[index] = localMarker
-			}
-		}
-	}
-	return recomputeStopEligible(status)
-}
-
-func recomputeStopEligible(status eruncommon.EnvironmentIdleStatus) eruncommon.EnvironmentIdleStatus {
-	if !status.ManagedCloud {
-		status.StopEligible = false
-		if status.StopBlockedReason == "" {
-			status.StopBlockedReason = "environment is not cloud-managed"
-		}
-		return status
-	}
-	if status.OutsideWorkingHours {
-		status.StopEligible = true
-		status.StopBlockedReason = ""
-		return status
-	}
-	for _, marker := range status.Markers {
-		if marker.Name == "working-hours" {
-			continue
-		}
-		if !marker.Idle {
-			status.StopEligible = false
-			status.StopBlockedReason = uiStopBlockedReason(status.Markers)
-			return status
-		}
-	}
-	status.StopEligible = true
-	status.StopBlockedReason = ""
-	return status
-}
-
-func uiStopBlockedReason(markers []eruncommon.EnvironmentIdleMarker) string {
-	for _, marker := range markers {
-		if marker.Name == "working-hours" || marker.Idle {
-			continue
-		}
-		name := strings.TrimSpace(marker.Name)
-		reason := strings.TrimSpace(marker.Reason)
-		if name == "" {
-			return reason
-		}
-		if reason == "" {
-			return name
-		}
-		return name + ": " + reason
-	}
-	return ""
-}
-
-func (a *App) maybeStopIdleCloudEnvironment(result eruncommon.OpenResult, status eruncommon.EnvironmentIdleStatus) {
-	if !status.ManagedCloud || !status.StopEligible {
-		return
-	}
-	cloudContext, ok, err := a.linkedCloudContext(result.EnvConfig)
-	if err != nil || !ok {
-		return
-	}
-	key := selectionKey(uiSelection{Tenant: result.Tenant, Environment: result.Environment})
-	busyKey := environmentBusyKey(uiSelection{Tenant: result.Tenant, Environment: result.Environment})
-	a.mu.Lock()
-	if a.busyEnvs[busyKey] > 0 {
-		a.mu.Unlock()
-		return
-	}
-	if _, exists := a.idleStops[key]; exists {
-		a.mu.Unlock()
-		return
-	}
-	a.idleStops[key] = struct{}{}
-	a.mu.Unlock()
-
-	a.emitAppStatus(fmt.Sprintf("Stopping idle cloud context %s...", cloudContextDisplayName(cloudContext)), true)
-	go func() {
-		ctx := a.ctx
-		if ctx == nil {
-			ctx = context.Background()
-		}
-		if _, err := a.deps.stopCloudContext(ctx, cloudContext.Name); err != nil {
-			a.mu.Lock()
-			delete(a.idleStops, key)
-			a.mu.Unlock()
-			a.emitAppStatus(fmt.Sprintf("Failed to stop idle cloud context %s: %s", cloudContextDisplayName(cloudContext), err.Error()), false)
-			return
-		}
-		a.emitAppStatus(fmt.Sprintf("Stopped idle cloud context %s.", cloudContextDisplayName(cloudContext)), false)
-	}()
-}
-
-func activitySecondsUntilIdle(status eruncommon.EnvironmentIdleStatus) int64 {
-	var seconds int64
-	for _, marker := range status.Markers {
-		if marker.Name == "working-hours" || marker.Idle {
-			continue
-		}
-		if marker.SecondsRemaining > seconds {
-			seconds = marker.SecondsRemaining
-		}
-	}
-	return seconds
 }
 
 func (a *App) ensureMCPAvailable(ctx context.Context, result eruncommon.OpenResult) error {
